I regretted, I love to play, you should know that
Looks like now I make it! Well, you're right, the code sucks yet, but at
least we can see it working, it really needs a refactoring specially in
the Factory and ValidationException (I'm sorry, mom...), I was coding
like a monkey! That's because I took the base from Proposal 1, so I had
to change a lot of things with no precise idea of how (or if) it was
going to work.

Not so happy, though, the only way to define properties to exceptions
from rules is defining public properties like Validation does today.
Also, with this approach we don't have access to the Factory defined on
Validator object, but that's live, sometimes you win, sometimes you lose.

The very good point, compared to the Proposal 1, is that it doesn't need
to create one Context object every time a rule is applied. I can say I
like this proposal, yes, I do like this proposal, despite I think it's
not so fancy as proposal 1, perhaps because that one is my little child.
Any way, this proposal is really good!

Now I need to talk to the guys, I hope they can help me with that, it's
a lot of things to do, many changes... I cannot do it alone, guys,
please, help me, please! I buy you guys I beer, how about it? I know you
guys like beer. Think about it :D